Fixed bug causing erronous intersection. Now using ray/triangle-intersection in accordance to Real-Time Rendering.# This is a combination of 3 commits. Made an assortment of minor fixes and enhancements. Discovered erronous intersectional detection.
Solved intersection bug with new collission algorithm in accordance to Real-Time Rendering. Encountered new, albeit interesting, texturing bug in exchange.

Fixed erronous length-calculation of intersection vector.

+++ src/dv2520/Win.cpp
@@ -150,8 +150,8 @@ void Win::wWinKeyProc( WPARAM wParam ) {
 	}
 }
 void Win::wWinMouseProc( WPARAM wParam, LPARAM lParam ) {
-	float x = GET_X_LPARAM( lParam );
-	float y = GET_Y_LPARAM( lParam );
+	float x = (float)GET_X_LPARAM( lParam );
+	float y = (float)GET_Y_LPARAM( lParam );
 	m_inputQueue.mousePush( InputMouse( x, y ) );
 	
 	//if( ( wParam & MK_LBUTTON )!=0 ) {
